# docs/openapi.yml
openapi: 3.0.0
info:
  title: DashFrame API
  version: 0.1.0
  description: |
    ## Authentication

    This API uses tokens of an unknown type.
    The token is in the `Authorization` header of the request.  

    Example:
    ```
    Authorization: gho_pJ9dGXVKpfzZp4PUHSxYEq9hjk0h288Gwj4S
    ```

    ## Typical usage:

    Once logged-in, the client will fetch the widgets on the board:
    ```
    GET $API_URL/widgets
    ```

    Then it will listen for widget update events by starting a WebSocket connection:
    ```
    GET $API_URL/ws/events
    ```

    In the meantime, the client will fetch the list of available widget and services:
    ```
    GET $API_URL/services
    ```
servers:
  - url: http://localhost:8080/api/{apiVersion}
    description: Development API Server
    variables:
      apiVersion:
        default: v1
components:
  schemas:
...
    EpicGamesFreeGamesWidgetConfiguration:
      allOf:
        - $ref: "#/components/schemas/BaseWidgetConfiguration"
    EpicGamesFriendsListWidgetConfiguration:
      allOf:
        - $ref: "#/components/schemas/BaseWidgetConfiguration"
        - type: object
          description: Configuration for the Epic Games friends list widget
          properties:
            parameters:
              type: object
              properties:
                showOffline:
                  type: integer
                  enum: [0, 1]
              required:
                - showOffline
          required:
            - parameters
...
    Identifier:
      type: integer
      description: A resource identifier
      minimum: 0
    OptionalUserIdentifier:
      type: object
      description: The ID of an user, assumed to be the current user if not specified
      properties:
        userId:
          $ref: "#/components/schemas/Identifier"
    RssFeedWidgetConfiguration:
      allOf:
        - $ref: "#/components/schemas/BaseWidgetConfiguration"
        - type: object
          description: Configuration for the RSS feed widget
          properties:
            parameters:
              type: object
              properties:
                url:
                  type: string
                  format: uri
              required:
                - url
          required:
            - parameters
...
    ServiceDescription:
      type: object
      properties:
        name:
          type: string
          description: Indicates the name of the Service
          example: weather
        iconUrl:
          type: string
          format: uri
          description: The image URL of the service
        description:
          type: string
          description: Describes the service
        widgets:
          type: array
          description: Indicates the list of Widgets supported by this Service
          items:
            $ref: "#/components/schemas/WidgetDescription"
      required:
        - name
        - widgets
    ServiceInstance:
      type: object
      description: A service instance, a user may possess multiple instances (accounts) for any given service
      example:
        {
          "id": 24,
          "type": "steam",
          "username": "Xx_DarkGam3R_xX",
          "token": "c3VwZXIgc2VjcmV0IHRva2VuCg==",
          "createdAt": "2022-11-07 15:17:56.69100",
        }
      properties:
        id:
          $ref: "#/components/schemas/Identifier"
        type:
          $ref: "#/components/schemas/ServiceType"
        username:
          type: string
          description: An username/email, unique accross service instances of a user
        token:
          type: string
          description: Login token for the service
        createdAt:
          type: string
          format: date-time
          description: Created date-time
      required:
        - id
        - type
...
    ServiceType:
      description: "Service unique identifier"
      enum: ["epic_games", "rss", "steam", "twitch", "weather_time", "youtube"]
    SteamFriendsListWidgetConfiguration:
      allOf:
        - $ref: "#/components/schemas/BaseWidgetConfiguration"
        - type: object
          description: Configuration for the Steam friends list widget
          properties:
            parameters:
              type: object
              properties:
                showOffline:
                  type: integer
                  enum: [0, 1]
              required:
                - showOffline
          required:
            - parameters
    SteamGameNewsWidgetConfiguration:
      allOf:
        - $ref: "#/components/schemas/BaseWidgetConfiguration"
        - type: object
          description: Configuration for the Steam game news widget
          properties:
            parameters:
              type: object
              properties:
                gameNameOrId:
                  type: string
                  minLength: 1
              required:
                - gameNameOrId
          required:
            - parameters
    SteamGamePopulationWidgetConfiguration:
      allOf:
        - $ref: "#/components/schemas/BaseWidgetConfiguration"
        - type: object
          description: Configuration for the Steam game population widget
          properties:
            parameters:
              type: object
              properties:
                gameNameOrId:
                  type: string
                  minLength: 1
              required:
                - gameNameOrId
          required:
            - parameters
    TwitchFollowedChannelsWidgetConfiguration:
      allOf:
        - $ref: "#/components/schemas/BaseWidgetConfiguration"
        - type: object
          description: Configuration for the Twitch list of followed channels
          properties:
            parameters:
              type: object
              properties:
                filter:
                  type: string
                  enum: ["ALL", "STREAMING_ONLY"]
              required:
                - filter
          required:
            - parameters
...
    WeatherTimeWidgetConfiguration:
      allOf:
        - $ref: "#/components/schemas/BaseWidgetConfiguration"
        - type: object
          description: Configuration for the weather & time widget
          properties:
            parameters:
              type: object
              properties:
                location:
                  description: City name or address
                  type: string
              required:
                - location
          required:
            - parameters
    WidgetConfiguration:
      oneOf:
        - $ref: "#/components/schemas/EpicGamesFreeGamesWidgetConfiguration"
        - $ref: "#/components/schemas/EpicGamesFriendsListWidgetConfiguration"
        - $ref: "#/components/schemas/RssFeedWidgetConfiguration"
        - $ref: "#/components/schemas/SteamFriendsListWidgetConfiguration"
        - $ref: "#/components/schemas/SteamGameNewsWidgetConfiguration"
        - $ref: "#/components/schemas/SteamGamePopulationWidgetConfiguration"
        - $ref: "#/components/schemas/TwitchFollowedChannelsWidgetConfiguration"
        - $ref: "#/components/schemas/WeatherTimeWidgetConfiguration"
        - $ref: "#/components/schemas/YouTubeChannelStatisticsWidgetConfiguration"
        - $ref: "#/components/schemas/YouTubeSubscribedChannelsWidgetConfiguration"
        - $ref: "#/components/schemas/YouTubeVideoStatisticsWidgetConfiguration"
      discriminator:
        propertyName: type
        mapping:
          epic_games/free_games: "#/components/schemas/EpicGamesFreeGamesWidgetConfiguration"
          epic_games/friends_list: "#/components/schemas/EpicGamesFriendsListWidgetConfiguration"
          rss/feed: "#/components/schemas/RssFeedWidgetConfiguration"
          steam/friends_list: "#/components/schemas/SteamFriendsListWidgetConfiguration"
          steam/game_news: "#/components/schemas/SteamGameNewsWidgetConfiguration"
          steam/game_population: "#/components/schemas/SteamGamePopulationWidgetConfiguration"
          twitch/followed_channels: "#/components/schemas/TwitchFollowedChannelsWidgetConfiguration"
          weather_time/default: "#/components/schemas/WeatherTimeWidgetConfiguration"
          youtube/channel_statistics: "#/components/schemas/YouTubeChannelStatisticsWidgetConfiguration"
          youtube/subsribed_channels: "#/components/schemas/YouTubeSubscribedChannelsWidgetConfiguration"
          youtube/video_statistics: "#/components/schemas/YouTubeVideoStatisticsWidgetConfiguration"
...
    WidgetDescription:
      type: object
      properties:
        name:
          type: string
          description: Indicates the identifier of this Widget
          example: city_temperature
        description:
          type: string
          example: Display temperature for a city
        params:
          type: array
          description: Indicates the list of parameters to configure this Widget
          items:
            $ref: "#/components/schemas/WidgetParameterDescription"
      required:
        - name
        - description
        - params
...
    WidgetInstance:
      type: object
      example:
        {
          "id": 42,
          "service_id": 1,
          "order": 3,
          "column_pos": 1,
          "config":
            {
              "type": "steam/friends_list",
              "parameters": { "showOffline": 1 },
            },
          "createdAt": "2022-11-07 15:17:56.69100",
        }
      properties:
        id:
          $ref: "#/components/schemas/Identifier"
        name:
          description: User-defined name of this widget component, optional
          type: string
        serviceId:
          $ref: "#/components/schemas/Identifier"
        order:
          type: integer
          description: The ordering of this widget in the column
          minimum: 0
        columnPos:
          type: integer
          description: The column id of the widget
          minimum: 0
        config:
          $ref: "#/components/schemas/WidgetConfiguration"
        createdAt:
          type: string
          format: date-time
          description: Created date-time
      required:
        - id
        - service_id
        - order
        - column_pos
...
    WrappedApiError:
      type: object
      description: |
        An API error object wrapped as an object with an "error" key.
        Used to distinguish between different types of values
      example:
        {
          "error":
            {
              "code": 3,
              "message": 'Missing string parameter "config"."name"',
              "messageKey": "error.missing_parameter",
              "messageParameters":
                { "path": '"config"."name"', "type": "string" },
            },
        }
      properties:
        error:
          $ref: "#/components/schemas/ApiError"
      required:
        - error
    YouTubeChannelStatisticsWidgetConfiguration:
      allOf:
        - $ref: "#/components/schemas/BaseWidgetConfiguration"
        - type: object
          description: Configuration for the YouTube channel stats widget
          properties:
            parameters:
              type: object
              properties:
                channel:
                  type: string
                  minLength: 1
              required:
                - channel
          required:
            - parameters
    YouTubeSubscribedChannelsWidgetConfiguration:
      allOf:
        - $ref: "#/components/schemas/BaseWidgetConfiguration"
        - type: object
          description: Configuration for the YouTube list of subscribed channels
          properties:
            parameters:
              type: object
              properties:
                filter:
                  type: string
                  enum: ["ALL", "STREAMING_ONLY"]
              required:
                - filter
          required:
            - parameters
...
  responses:
    ErrorResponse:
      description: |
        Error HTTP response.

        Status codes:
        - 400: Returned if the requested was malformed
        - 401: The user is not logged in
        - 404: Returned if the requested resource doesn't exist, or the user does not have access 
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/WrappedApiError"
  parameters:
    optionalUserId:
      name: userId
      in: query
      required: false
      description: The ID of an user, assumed to be the current user if not specified
      schema:
        $ref: "#/components/schemas/Identifier"
    widgetId:
      name: widgetId
      in: path
      required: true
      description: "The ID of a widget"
      schema:
        $ref: "#/components/schemas/Identifier"
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
...
paths:
...
  /services:
    get:
      operationId: list_services
      summary: Returns the list of available services and services instances registered by the user
      security:
        - apiKey: []
      tags:
        - services
      parameters:
        - $ref: "#/components/parameters/optionalUserId"
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  available:
                    type: array
                    description: Indicates the list of Services supported by the server
                    items:
                      $ref: "#/components/schemas/ServiceDescription"
                  instances:
                    type: array
                    items:
                      $ref: "#/components/schemas/ServiceInstance"
                required:
                  - available
                  - instances
        "4XX":
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: create_service
      summary: Create a new service instance
      security:
        - apiKey: []
      tags:
        - services
      requestBody:
        content:
          application/json:
            example:
              {
                "type": "steam",
                "token": "c3VwZXIgc2VjcmV0IHRva2VuCg==",
                "username": "Xx_DarkGam3R_xX",
              }
            schema:
              type: object
              allOf:
                - $ref: "#/components/schemas/OptionalUserIdentifier"
                - type: object
                  properties:
                    type:
                      $ref: "#/components/schemas/ServiceType"
                    token:
                      type: string
                      description: The token returned by the external service, optional if not needed
                    username:
                      type: string
                      description: A string that identifies this specific account, not needed if the services does not have accounts
                  required:
                    - type
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceInstance"
        "4XX":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: delete_services
      summary: Removes all registered service instances of the user
      security:
        - apiKey: []
      tags:
        - services
      parameters:
        - $ref: "#/components/parameters/optionalUserId"
      responses:
        "204":
          description: All services were deleted successfully
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /services/{serviceId}:
    delete:
      operationId: delete_service
      summary: "Delete a specific service instance"
      security:
        - apiKey: []
      tags:
        - services
      parameters:
        - name: serviceId
          required: true
          in: path
          schema:
            $ref: "#/components/schemas/Identifier"
      responses:
        "204":
          description: All services of type or service was deleted successfully
        "4XX":
          $ref: "#/components/responses/ErrorResponse"
  /widgets:
    get:
      operationId: list_widgets
      summary: Returns the layout of the requested user's board
      security:
        - apiKey: []
      tags:
        - widgets
      parameters:
        - $ref: "#/components/parameters/optionalUserId"
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WidgetInstance"
        "4XX":
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: create_widget
      summary: Create a new widget instance in the given user's board
      security:
        - apiKey: []
      tags:
        - widgets
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/OptionalUserIdentifier"
                - type: object
                  properties:
                    serviceId:
                      $ref: "#/components/schemas/Identifier"
                    config:
                      $ref: "#/components/schemas/WidgetConfiguration"
                  required:
                    - serviceId
                    - config
      responses:
        "201":
          description: Widget was created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetInstance"
        "4XX":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: delete_widgets
      summary: Clear all the widgets in the user's board
      security:
        - apiKey: []
      tags:
        - widgets
      parameters:
        - $ref: "#/components/parameters/optionalUserId"
      responses:
        "204":
          description: Board was cleared successfully
        "4XX":
          $ref: "#/components/responses/ErrorResponse"
  /widgets/{widgetId}:
    get:
      operationId: get_widget
      summary: Get information about a specific widget
      security:
        - apiKey: []
      tags:
        - widgets
      parameters:
        - $ref: "#/components/parameters/widgetId"
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetInstance"
        "4XX":
          $ref: "#/components/responses/ErrorResponse"
    patch:
      operationId: update_widget
      summary: Update an existing widget's configuration
      security:
        - apiKey: []
      tags:
        - widgets
      parameters:
        - $ref: "#/components/parameters/widgetId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                config:
                  $ref: "#/components/schemas/WidgetConfiguration"
              required:
                - config
      responses:
        "204":
          description: The widget was updated successfully
        "4XX":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: delete_widget
      summary: Delete a widget
      security:
        - apiKey: []
      tags:
        - widgets
      parameters:
        - $ref: "#/components/parameters/widgetId"
      responses:
        "204":
          description: The widget was deleted successfully
        "4XX":
          $ref: "#/components/responses/ErrorResponse"
...